+++ src/pygaps/calculations/models_isotherm/jensenseaton.py
@@ -129,7 +129,7 @@ class JensenSeaton(IsothermModel):
         float
             Spreading pressure at specified pressure.
         """
-        return NotImplementedError
+        return scipy.integrate.quad(lambda x: self.loading(x) / x, 0, pressure)[0]
 
     def default_guess(self, data, loading_key, pressure_key):
         """
@@ -154,5 +154,5 @@ class JensenSeaton(IsothermModel):
 
         return {"KH": saturation_loading * langmuir_k,
                 "a": 1,
-                "b": 0,
+                "b": 1,
                 "c": 1, }